Splitting up login into more files.
