+++ database/Tests-Sign-in.py
@@ -1,0 +1,12 @@
+'''
+username = input("Please enter a username: ")
+firstname = input("Please enter your first name: ")
+middlename = input("Please enter your middle name: ")
+lastname = input("Please enter your last name: ")
+email = input("Please enter your email: ")
+password = input("Please enter your password: ")
+password2 = input("Please enter your password: ")
+
+   username = input("Please enter a username: ")
+    password = input("Please enter your password: ")
+'''
